--- diploma-libs/slide-raster-graphics/CommandStorage.js
@@ -1,42 +1,0 @@
-// CommandStorage.js
-
-export class CommandStorage {
-    constructor() {
-        this.storage = {}; // объект для хранения команд по ключам
-    }
-
-    // Метод для сохранения команд
-    saveCommands(key, commands) {
-        if (!Array.isArray(commands)) {
-            throw new Error('Commands must be an array');
-        }
-        this.storage[key] = commands;
-        console.log(`Commands saved under key: ${key}`);
-    }
-
-    // Метод для получения команд по ключу
-    getCommands(key) {
-        const commands = this.storage[key];
-        if (!commands) {
-            console.warn(`No commands found for key: ${key}`);
-            return null;
-        }
-        return commands;
-    }
-
-    // Метод для удаления команд по ключу
-    removeCommands(key) {
-        if (this.storage[key]) {
-            delete this.storage[key];
-            console.log(`Commands removed for key: ${key}`);
-        } else {
-            console.warn(`No commands found for key: ${key}`);
-        }
-    }
-
-    // Метод для получения всех доступных ключей
-    getAllKeys() {
-        return Object.keys(this.storage);
-    }
-
-}
